minor improvements

+++ Nuvola/Plugins/dump.py
@@ -28,6 +28,8 @@ async def dump_to_file(client: Nuvola, chat: int | str, filename: str):
 # Dump command
 @Nuvola.on_message(filters.me & filters.command("dump", PREFIX))
 async def dump_cmd(client: Nuvola, message: Message):
+    # Initializing chat_to_get
+    chat_to_get = None
     # Check whether an argument is provided or not
     if (len(message.command) == 2):
         # If yes, the script will get chat info based on that arg
@@ -37,25 +39,27 @@ async def dump_cmd(client: Nuvola, message: Message):
         # If the message is sent in a public chat, the script will get its info
         if (message.chat.type != enums.ChatType.PRIVATE):
             chat_to_get = message.chat.id
-        else:
-            # The command must be used in a public chat or providing an arg
-            await message.edit_text(ARG_MISSING)
 
-    # Get chat info
-    chat = await client.get_chat(chat_to_get)
-    # Get all chat members and write the scraped data to a file
-    file = await dump_to_file(client, chat=chat.id, filename=chat.title)
-    # Id or username of the chat where you want the dump file to be sent, "me" for saved messages
-    destination = "me"
-    # Caption of the dump file
-    caption = f"🆔 {chat.id}\n💬 {chat.invite_link}"
-    # Send dump file using provided arguments
-    await client.send_document(chat_id=destination, file_name=file, caption=caption)
-    # Free memory
-    os.remove(file)
+    # If chat_to_get is not None, get chat info, dump members, and send the file
+    if chat_to_get:
+        # Get chat info
+        chat = await client.get_chat(chat_to_get)
+        # Get all chat members and write the scraped data to a file
+        file = await dump_to_file(client, chat=chat.id, filename=chat.title)
+        # Id or username of the chat where you want the dump file to be sent, "me" for saved messages
+        destination = "me"
+        # Caption of the dump file
+        caption = f"∙ ID » {chat.id}\n∙ Username » @{chat.username}"
+        # Send dump file using provided arguments
+        await client.send_document(chat_id=destination, document=file, caption=caption)
+        # Free memory
+        os.remove(file)
+        # Edit message
+        await message.edit_text("Dump successful, check your saved messages.")
+    # If chat_to_get is None it means that no arguments where provided
+    else:
+        await message.edit_text(ARG_MISSING)
 
-    # Edit message
-    await message.edit_text("Dump successful, check your saved messages.")
     # Wait x seconds
     await asyncio.sleep(2)
     # Delete message
